% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ims}
\alias{plot_ims}
\title{Plot images from target 'result'}
\usage{
plot_ims(nmax = 24)
}
\arguments{
\item{nmax}{maximum number of images to plot in one frame (default: 24)}
}
\description{
A faceted frame is set up and each image is plotted.
}
