% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ims}
\alias{plot_ims}
\title{Plot images from target 'result'}
\usage{
plot_ims(nmax = 24)
}
\arguments{
\item{nmax}{maximum number of images to plot in one frame (default: 24)}
}
\description{
A faceted frame is set up and each image is plotted.
}
\details{
This is for interactive use, i.e. \code{plot_ims(nmax = 12)} it can't run as a target.
}
